Extract gradle logic
