+++ slack-plugin/src/main/kotlin/slack/gradle/bazel/JvmProjectSpec.kt
@@ -18,8 +18,6 @@ package slack.gradle.bazel
 import com.grab.grazel.bazel.starlark.BazelDependency
 import com.grab.grazel.bazel.starlark.asString
 import com.grab.grazel.bazel.starlark.statements
-import java.io.File
-import java.util.SortedSet
 import okio.FileSystem
 import okio.Path
 import okio.Path.Companion.toOkioPath
@@ -27,19 +25,8 @@ import org.gradle.api.DefaultTask
 import org.gradle.api.NamedDomainObjectProvider
 import org.gradle.api.Project
 import org.gradle.api.artifacts.ResolvableConfiguration
-import org.gradle.api.artifacts.component.ComponentArtifactIdentifier
-import org.gradle.api.artifacts.component.ProjectComponentIdentifier
-import org.gradle.api.file.RegularFileProperty
-import org.gradle.api.provider.Property
-import org.gradle.api.provider.Provider
-import org.gradle.api.provider.SetProperty
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.Optional
-import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.UntrackedTask
-import org.gradle.internal.component.external.model.ModuleComponentArtifactIdentifier
-import org.gradle.internal.component.local.model.PublishArtifactLocalArtifactMetadata
 import slack.gradle.SlackExtension
 import slack.gradle.SlackProperties
 
@@ -143,27 +130,7 @@ internal class JvmProjectSpec(builder: Builder) :
 }
 
 @UntrackedTask(because = "Generates a Bazel BUILD file for a Kotlin JVM project")
-internal abstract class JvmProjectBazelTask : DefaultTask() {
-  @get:Input abstract val targetName: Property<String>
-  @get:Input abstract val ruleSource: Property<String>
-
-  @get:Input abstract val projectDir: Property<File>
-
-  @get:Input abstract val deps: SetProperty<ComponentArtifactIdentifier>
-  @get:Input abstract val exportedDeps: SetProperty<ComponentArtifactIdentifier>
-  @get:Input abstract val testDeps: SetProperty<ComponentArtifactIdentifier>
-  @get:Input abstract val kspDeps: SetProperty<ComponentArtifactIdentifier>
-  @get:Input abstract val kaptDeps: SetProperty<ComponentArtifactIdentifier>
-  @get:Input abstract val compilerPlugins: SetProperty<Dep>
-  @get:Input abstract val kspProcessors: SetProperty<KspProcessor>
-
-  // Features
-  @get:Optional @get:Input abstract val moshix: Property<Boolean>
-  @get:Optional @get:Input abstract val redacted: Property<Boolean>
-  @get:Optional @get:Input abstract val parcelize: Property<Boolean>
-  @get:Optional @get:Input abstract val autoService: Property<Boolean>
-
-  @get:OutputFile abstract val outputFile: RegularFileProperty
+internal abstract class JvmProjectBazelTask : DefaultTask(), CommonJvmProjectBazelTask {
 
   init {
     group = "bazel"
@@ -172,94 +139,10 @@ internal abstract class JvmProjectBazelTask : DefaultTask() {
 
   @TaskAction
   fun generate() {
-    val deps = deps.mapDeps()
-    val exportedDeps = exportedDeps.mapDeps()
-    val testDeps = testDeps.mapDeps()
-
-    // Only moshix and redacted are supported in JVM projects
-    val compilerPlugins = compilerPlugins.get().toMutableList()
-
-    val kspProcessors = kspProcessors.get().toMutableList()
-
-    if (moshix.getOrElse(false)) {
-      // TODO we technically could choose IR or KSP for this, but for now assume IR
-      compilerPlugins.add(CompilerPluginDeps.moshix)
-      kspProcessors.add(KspProcessors.moshiProguardRuleGen)
-    }
-    if (redacted.getOrElse(false)) {
-      compilerPlugins.add(CompilerPluginDeps.redacted)
-    }
-    if (parcelize.getOrElse(false)) {
-      compilerPlugins.add(CompilerPluginDeps.parcelize)
-    }
-    if (autoService.getOrElse(false)) {
-      kspProcessors.add(KspProcessors.autoService)
-    }
-
-    // TODO make this pluggable and single-pass
-    val mappedKspDeps = kspDeps.mapDeps()
-    when {
-      mappedKspDeps.any { "guinness" in it.toString() } -> {
-        logger.lifecycle("[KSP] Adding guinness compiler")
-        kspProcessors += KspProcessors.guinness
-      }
-      mappedKspDeps.any { "feature-flag/compiler" in it.toString() } -> {
-        logger.lifecycle("[KSP] Adding feature flag compiler")
-        kspProcessors += KspProcessors.featureFlag
-      }
-    }
-    val allKspDeps = mappedKspDeps.map { it.toString() }
-
-    // TODO kapt
-
     JvmProjectSpec.Builder(targetName.get())
-      .apply {
-        this@JvmProjectBazelTask.ruleSource.orNull?.let(::ruleSource)
-        deps.forEach { addDep(it) }
-        exportedDeps.forEach { addExportedDep(it) }
-        testDeps.forEach { addTestDep(it) }
-        compilerPlugins.forEach { addCompilerPlugin(it) }
-        kspProcessors.forEach { addKspProcessor(it.withAddedDeps(allKspDeps)) }
-      }
+      .applyCommonJvmConfig()
       .build()
       .writeTo(outputFile.asFile.get().toOkioPath())
-  }
-
-  private fun SetProperty<ComponentArtifactIdentifier>.mapDeps(): SortedSet<Dep> {
-    return map { result ->
-        result.asSequence().mapNotNull { component ->
-          when (component) {
-            is ModuleComponentArtifactIdentifier -> {
-              val componentId = component.componentIdentifier
-              val identifier = "${componentId.group}:${componentId.module}"
-              Dep.Remote.fromMavenIdentifier(identifier)
-            }
-            is PublishArtifactLocalArtifactMetadata -> {
-              val projectIdentifier = component.componentIdentifier
-              check(projectIdentifier is ProjectComponentIdentifier)
-              // Map to "path/to/local/dependency1" format
-              Dep.Local(
-                projectIdentifier.projectPath.removePrefix(":").replace(":", "/"),
-                target = "lib",
-              )
-            }
-            else -> {
-              System.err.println("Unknown component type: $component (${component.javaClass})")
-              null
-            }
-          }
-        }
-      }
-      .get()
-      .toSortedSet()
-  }
-
-  protected fun resolvedDependenciesFrom(
-    provider: NamedDomainObjectProvider<ResolvableConfiguration>
-  ): Provider<List<ComponentArtifactIdentifier>> {
-    return provider.flatMap { configuration ->
-      configuration.incoming.artifacts.resolvedArtifacts.map { it.map { it.id } }
-    }
   }
 
   companion object {
@@ -276,19 +159,16 @@ internal abstract class JvmProjectBazelTask : DefaultTask() {
       project.tasks
         .register("generateBazel", JvmProjectBazelTask::class.java, slackProperties)
         .configure {
-          targetName.set(project.name)
-          ruleSource.set(slackProperties.bazelRuleSource)
-          projectDir.set(project.layout.projectDirectory.asFile)
-          deps.set(resolvedDependenciesFrom(depsConfiguration))
-          exportedDeps.set(resolvedDependenciesFrom(exportedDepsConfiguration))
-          testDeps.set(resolvedDependenciesFrom(testConfiguration))
-          kspConfiguration?.let { kspDeps.set(resolvedDependenciesFrom(it)) }
-          kaptConfiguration?.let { kaptDeps.set(resolvedDependenciesFrom(it)) }
-          outputFile.set(project.layout.projectDirectory.file("BUILD.bazel"))
-          moshix.set(slackExtension.featuresHandler.moshiHandler.moshiCodegen)
-          redacted.set(slackExtension.featuresHandler.redacted)
-          parcelize.set(project.pluginManager.hasPlugin("org.jetbrains.kotlin.plugin.parcelize"))
-          autoService.set(slackExtension.featuresHandler.autoService)
+          configureCommonJvm(
+            project,
+            slackProperties,
+            depsConfiguration,
+            exportedDepsConfiguration,
+            testConfiguration,
+            kspConfiguration,
+            kaptConfiguration,
+            slackExtension,
+          )
         }
     }
   }
